--- packages/bluesnap-direct-integration/src/bluesnap-direct-apms/create-bluesnap-direct-ecp-payment-strategy.ts
@@ -1,14 +1,0 @@
-import {
-    PaymentStrategyFactory,
-    toResolvableModule,
-} from '@bigcommerce/checkout-sdk/payment-integration-api';
-
-import BlueSnapDirectEcpPaymentStrategy from './bluesnap-direct-ecp-payment-strategy';
-
-const createBlueSnapDirectEcpPaymentStrategy: PaymentStrategyFactory<
-    BlueSnapDirectEcpPaymentStrategy
-> = (paymentIntegrationService) => new BlueSnapDirectEcpPaymentStrategy(paymentIntegrationService);
-
-export default toResolvableModule(createBlueSnapDirectEcpPaymentStrategy, [
-    { id: 'ecp', gateway: 'bluesnapdirect' },
-]);
