+++ packages/bluesnap-direct-integration/src/index.ts
@@ -1,8 +1,6 @@
 export { default as createBlueSnapDirectCreditCardPaymentStrategy } from './bluesnap-direct-credit-card/create-bluesnap-direct-credit-card-payment-strategy';
-export { default as createBlueSnapDirectEcpPaymentStrategy } from './bluesnap-direct-apms/create-bluesnap-direct-ecp-payment-strategy';
-export { default as createBlueSnapDirectAPMPaymentStrategy } from './bluesnap-direct-apms/create-bluesnap-direct-apm-payment-strategy';
-export { default as createBlueSnapDirectSepaPaymentStrategy } from './bluesnap-direct-apms/create-bluesnap-direct-sepa-payment-strategy';
-export { default as createBlueSnapDirectIdealPaymentStrategy } from './bluesnap-direct-apms/create-bluesnap-direct-ideal-payment-strategy';
+export { default as createBlueSnapDirectAPMPaymentStrategy } from './bluesnap-direct-apm/create-bluesnap-direct-apm-payment-strategy';
+
 export {
     BlueSnapDirectAPMInitializeOptions,
     WithBlueSnapDirectAPMPaymentInitializeOptions,
