+++ packages/core/src/payment/index.ts
@@ -21,7 +21,7 @@ export { default as PaymentActionCreator } from './payment-action-creator';
 export {
     default as Payment,
     CreditCardInstrument,
-    IdealInstrument,
+    WithIdealInstrument,
     WithCheckoutcomFawryInstrument,
     WithCheckoutcomSEPAInstrument,
     HostedCreditCardInstrument,
