+++ packages/core/src/payment/strategies/checkoutcom-custom/checkoutcom-ideal/checkoutcom-ideal-payment-strategy.ts
@@ -1,7 +1,7 @@
 import { InternalCheckoutSelectors } from '../../../../checkout';
 import { OrderRequestBody } from '../../../../order';
 import { PaymentArgumentInvalidError } from '../../../errors';
-import { IdealInstrument, PaymentInstrument } from '../../../payment';
+import { PaymentInstrument, WithIdealInstrument } from '../../../payment';
 import { PaymentRequestOptions } from '../../../payment-request-options';
 import CheckoutcomCustomPaymentStrategy from '../checkoutcom-custom-payment-strategy';
 
@@ -42,7 +42,7 @@ export default class CheckoutcomiDealPaymentStrategy extends CheckoutcomCustomPa
     private _createFormattedPayload(
         methodId: string,
         paymentData: PaymentInstrument,
-    ): IdealInstrument | undefined {
+    ): WithIdealInstrument | undefined {
         if (CHECKOUTCOM_IDEAL_PAYMENT_METHOD === methodId && 'bic' in paymentData) {
             return { bic: paymentData.bic };
         }
