+++ packages/payment-integration-api/src/index.ts
@@ -110,12 +110,12 @@ export {
     SubmitOrderAction,
 } from './order';
 export {
-    EcpInstrument,
+    WithEcpInstrument,
+    WithSepaInstrument,
+    WithIdealInstrument,
     BlueSnapDirectEcpPayload,
-    SepaInstrument,
-    IdealInstrument,
-    IdealPayload,
     BlueSnapDirectSepaPayload,
+    IdealPayload,
     CardInstrument,
     CreditCardInstrument,
     WithBankAccountInstrument,
@@ -132,7 +132,6 @@ export {
     PaymentMethod,
     PaymentRequestOptions,
     PaymentResponseBody,
-    AdditionalActionRequired,
     PaymentStatusTypes,
     PaymentStrategy,
     PaymentInstrument,
