+++ packages/payment-integration-api/src/payment/index.ts
@@ -3,14 +3,14 @@ export { AccountInstrument, CardInstrument } from './instrument';
 
 export {
     default as Payment,
-    EcpInstrument,
     IdealPayload,
     BlueSnapDirectEcpPayload,
-    SepaInstrument,
     BlueSnapDirectSepaPayload,
+    WithSepaInstrument,
+    WithEcpInstrument,
     CreditCardInstrument,
     WithBankAccountInstrument,
-    IdealInstrument,
+    WithIdealInstrument,
     WithCheckoutcomFawryInstrument,
     WithCheckoutcomSEPAInstrument,
     HostedCreditCardInstrument,
@@ -45,7 +45,7 @@ export { PaymentInitializeOptions } from './payment-initialize-options';
 export { PaymentRequestOptions } from './payment-request-options';
 export { default as PaymentMethod } from './payment-method';
 export { default as PaymentResponse } from './payment-response';
-export { default as PaymentResponseBody, AdditionalActionRequired } from './payment-response-body';
+export { default as PaymentResponseBody } from './payment-response-body';
 export { default as PaymentStrategy } from './payment-strategy';
 export { default as PaymentStrategyFactory } from './payment-strategy-factory';
 export { default as PaymentStrategyResolveId } from './payment-strategy-resolve-id';
